+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 ## 1.6.0
 
 * Added buttons for Open File, Open Changes, Change Base, Change Repository (if more than one workspace folder open), and Refresh (if auto refresh is disabled) [#46](https://github.com/letmaik/vscode-git-tree-compare/issues/46)
-* Scan direct subfolders of workspace folders for repositories [#48](https://github.com/letmaik/vscode-git-tree-compare/issues/48)
+* Scan for repositories in direct subfolders of workspace folders [#48](https://github.com/letmaik/vscode-git-tree-compare/issues/48)
+* Show full paths in tree item tooltips
 * Fixed various issues when adding/removing workspace folders
 
 ## 1.5.0